fix signup bug completely
* cleanup after signup bug hotfix
* add isInvited property to UserModel
